+++ Apps/AdminWebClient/src/AdminWebClient.csproj
@@ -37,6 +37,12 @@
     <Content Include="ClientApp\tsconfig.json" />
     <Content Include="Server\Controllers\AuthenticationController.cs" />
   </ItemGroup>
+  <ItemGroup>
+    <None Remove="Server\Delegates\" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Server\Delegates\" />
+  </ItemGroup>
   <Target Name="DebugEnsureNodeEnv" BeforeTargets="Build" Condition=" '$(Configuration)' == 'Debug' And !Exists('$(SpaRoot)node_modules') ">
     <!-- Ensure Node.js is installed -->
     <Exec Command="node --version" ContinueOnError="true">
