+++ Apps/AdminWebClient/src/Server/Delegates/RestImmunizationAdminDelegate.cs
@@ -1,0 +1,63 @@
+//-------------------------------------------------------------------------
+// Copyright © 2019 Province of British Columbia
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//-------------------------------------------------------------------------
+namespace HealthGateway.Admin.Server.Delegates
+{
+    using System;
+    using System.Threading.Tasks;
+    using HealthGateway.Admin.Models;
+    using HealthGateway.Common.Models;
+    using HealthGateway.Common.Models.PHSA;
+    using HealthGateway.Common.Services;
+    using Microsoft.AspNetCore.Http;
+    using Microsoft.Extensions.Configuration;
+    using Microsoft.Extensions.Logging;
+
+    /// <inheritdoc/>
+    public class RestImmunizationAdminDelegate : IImmunizationAdminDelegate
+    {
+        private const string ImmunizationConfigSectionKey = "Immunization";
+        private readonly ILogger logger;
+        private readonly IHttpClientService httpClientService;
+        private readonly ImmunizationConfig immunizationConfig;
+        private readonly IHttpContextAccessor httpContextAccessor;
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="RestImmunizationAdminDelegate"/> class.
+        /// </summary>
+        /// <param name="logger">Injected Logger Provider.</param>
+        /// <param name="httpClientService">The injected http client service.</param>
+        /// <param name="configuration">The injected configuration provider.</param>
+        /// <param name="httpContextAccessor">The Http Context accessor.</param>
+        public RestImmunizationAdminDelegate(
+            ILogger<RestImmunizationAdminDelegate> logger,
+            IHttpClientService httpClientService,
+            IConfiguration configuration,
+            IHttpContextAccessor httpContextAccessor)
+        {
+            this.logger = logger;
+            this.httpClientService = httpClientService;
+            this.httpContextAccessor = httpContextAccessor;
+            this.immunizationConfig = new ImmunizationConfig();
+            configuration.Bind(ImmunizationConfigSectionKey, this.immunizationConfig);
+        }
+
+        /// <inheritdoc/>
+        public Task<RequestResult<PHSAResult<ImmunizationResponse>>> GetImmunizations(string phn, DateTime birthDate, int pageIndex = 0)
+        {
+            throw new System.NotImplementedException();
+        }
+    }
+}
