+++ Apps/Common/src/Models/PHSA/PHSALoadState.cs
@@ -15,7 +15,6 @@
 // -------------------------------------------------------------------------
 namespace HealthGateway.Common.Models.PHSA
 {
-    using System.Collections.Generic;
     using System.Text.Json.Serialization;
 
     /// <summary>
