+++ Apps/Common/src/Models/PHSA/Recommendation/DateCriterionCode.cs
@@ -13,27 +13,19 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 // -------------------------------------------------------------------------
-namespace HealthGateway.Common.Models.PHSA
+namespace HealthGateway.Common.Models.PHSA.Recommendation
 {
     using System.Text.Json.Serialization;
 
     /// <summary>
-    /// Representation of the result sent by PHSA.
+    /// The PHSA Date Criterion Code data model.
     /// </summary>
-    /// <typeparam name="T">The result object type.</typeparam>
-    public class PHSAResult<T>
+    public class DateCriterionCode
     {
         /// <summary>
-        /// Gets or sets the LoadState.
+        /// Gets or sets the Date Criterion Code text.
         /// </summary>
-        [JsonPropertyName("loadState")]
-        public PHSALoadState LoadState { get; set; } = new PHSALoadState();
-
-        /// <summary>
-        /// Gets or sets the result section.
-        /// </summary>
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Usage", "CA2227:Collection properties should be read only", Justification = "Team decision")]
-        [JsonPropertyName("result")]
-        public T? Result { get; set; } = default(T);
+        [JsonPropertyName("text")]
+        public string Text { get; set; } = string.Empty;
     }
 }
