+++ Apps/Common/src/Models/PHSA/Recommendation/TargetDiseaseResponse.cs
@@ -13,7 +13,7 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 // -------------------------------------------------------------------------
-namespace HealthGateway.Immunization.Models.PHSA.Recommendation
+namespace HealthGateway.Common.Models.PHSA.Recommendation
 {
     using System.Collections.Generic;
     using System.Text.Json.Serialization;
