+++ Apps/Immunization/src/Services/ImmunizationService.cs
@@ -22,7 +22,6 @@ namespace HealthGateway.Immunization.Services
     using HealthGateway.Common.Models.PHSA;
     using HealthGateway.Immunization.Delegates;
     using HealthGateway.Immunization.Models;
-    using HealthGateway.Immunization.Models.PHSA;
     using HealthGateway.Immunization.Parser;
 
     /// <summary>
