+++ Apps/Immunization/test/unit/Delegates.Test/ImmsDelegateTests.cs
@@ -18,7 +18,6 @@ namespace HealthGateway.Immunization.Test.Delegates
     using System;
     using System.Collections.Generic;
     using System.Globalization;
-    using System.Linq;
     using System.Net;
     using System.Net.Http;
     using System.Security.Claims;
@@ -29,11 +28,9 @@ namespace HealthGateway.Immunization.Test.Delegates
     using HealthGateway.Common.ErrorHandling;
     using HealthGateway.Common.Models;
     using HealthGateway.Common.Models.PHSA;
+    using HealthGateway.Common.Models.PHSA.Recommendation;
     using HealthGateway.Common.Services;
     using HealthGateway.Immunization.Delegates;
-    using HealthGateway.Immunization.Models;
-    using HealthGateway.Immunization.Models.PHSA;
-    using HealthGateway.Immunization.Models.PHSA.Recommendation;
     using Microsoft.AspNetCore.Authentication;
     using Microsoft.AspNetCore.Authentication.JwtBearer;
     using Microsoft.AspNetCore.Http;
